Minor edits

+++ src/main/java/ru/sumarokov/premium_calculation/service/InsuranceService.java
@@ -3,6 +3,7 @@ package ru.sumarokov.premium_calculation.service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ru.sumarokov.premium_calculation.entity.Insurance;
+import ru.sumarokov.premium_calculation.exception.EntityNotFoundException;
 import ru.sumarokov.premium_calculation.repository.InsuranceRepository;
 
 import java.util.List;
@@ -23,7 +24,8 @@ public class InsuranceService {
     }
 
     public Insurance getInsurance(Long id) {
-        return insuranceRepository.findById(id).orElseThrow();
+        return insuranceRepository.findById(id)
+                .orElseThrow(EntityNotFoundException::new);
     }
 
     public void saveInsurance(Insurance insurance) {
@@ -31,6 +33,9 @@ public class InsuranceService {
     }
 
     public void deleteInsurance(Long id) {
+        if (!insuranceRepository.existsById(id)) {
+            throw new EntityNotFoundException();
+        }
         insuranceRepository.deleteById(id);
     }
 }
